# internal/transport/httpapi/openapi.yaml
openapi: 3.1.0
info:
  title: HCAI CHAT API
  version: 0.2.0
  description: Contract for the international-first HCAI CHAT platform.
servers:
  - url: /api/v1
...
paths:
...
  /account/profile:
    patch:
      tags: [Identity]
      operationId: updateProfile
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProfileUpdate' } } }
      responses:
        '200':
          description: Current profile updated with audit evidence
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties: { user: { $ref: '#/components/schemas/User' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
  /account/sessions:
    get:
      tags: [Identity]
      operationId: listAccountSessions
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: cursor, in: query, schema: { type: string }, description: Opaque stable account-session cursor }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        '200':
          description: Current account sessions with safe device and network hints
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountSessionPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /account/sessions/{sessionId}:
    delete:
      tags: [Identity]
      operationId: revokeAccountSession
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/SessionId' }]
      responses:
        '204': { description: Owned session revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /account/sessions/revoke-others:
    post:
      tags: [Identity]
      operationId: revokeOtherAccountSessions
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: All other active sessions revoked
          content:
            application/json:
              schema:
                type: object
                required: [revokedCount]
                properties: { revokedCount: { type: integer, minimum: 0 } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/oauth/{provider}/start:
    post:
      tags: [Identity]
      operationId: startOAuth
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [google, github] }
      responses:
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hcai_session
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: hcai_sk
  parameters:
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ValidationFailed:
      description: Request failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unavailable:
      description: Required provider unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ProfileUpdate:
      type: object
      additionalProperties: false
      required: [displayName, locale, timezone]
      properties:
        displayName: { type: string, minLength: 2, maxLength: 80 }
        locale: { type: string, enum: [en-US, zh-CN] }
        timezone: { type: string, maxLength: 80 }
...
    User:
      type: object
      required: [id, email, handle, displayName, role, status, locale, timezone, emailVerified, permissions]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        handle:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [member, creator, publisher, moderator, admin]
        status:
          type: string
          enum: [active, suspended, deleted]
        locale:
          type: string
        timezone:
          type: string
        emailVerified:
          type: boolean
        permissions:
          type: array
          items: { type: string }
...
    AccountSession:
      type: object
      required: [id, clientLabel, status, current, createdAt, lastSeenAt, expiresAt]
      properties:
        id: { type: string, format: uuid }
        clientLabel: { type: string }
        networkHint: { type: string }
        status: { type: string, enum: [active, revoked, expired] }
        current: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        revokedAt: { type: string, format: date-time }
    AccountSessionPage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AccountSession' } }
        nextCursor: { type: string }
...
    APIError:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        retryable:
          type: boolean
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/APIError'